requirements.txt
seaborn
pandas
matplotlib
numpy
platypus-opt